clearing pom.xml, custom runtime exception has been created for further inheritance of all bot exceptions from it
