--- src/main/java/dev/tssvett/schedule_bot/backend/exception/FacultyNotExistException.java
@@ -1,7 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class FacultyNotExistException extends RuntimeException {
-    public FacultyNotExistException(String message) {
-        super(message);
-    }
-}
