--- src/main/java/dev/tssvett/schedule_bot/backend/exception/GroupNotExistException.java
@@ -1,7 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class GroupNotExistException extends RuntimeException {
-    public GroupNotExistException(String message) {
-        super(message);
-    }
-}
