--- src/main/java/dev/tssvett/schedule_bot/backend/exception/NotValidRegistrationStateException.java
@@ -1,8 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class NotValidRegistrationStateException extends RuntimeException {
-
-    public NotValidRegistrationStateException(String message) {
-        super(message);
-    }
-}
