--- src/main/java/dev/tssvett/schedule_bot/backend/exception/ParseException.java
@@ -1,8 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class ParseException extends RuntimeException {
-
-    public ParseException(String message) {
-        super(message);
-    }
-}
