--- src/main/java/dev/tssvett/schedule_bot/backend/exception/PostBeanProcessorException.java
@@ -1,8 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class PostBeanProcessorException extends RuntimeException {
-
-    public PostBeanProcessorException(String message) {
-        super(message);
-    }
-}
