+++ src/main/java/dev/tssvett/schedule_bot/backend/exception/ScheduleBotException.java
@@ -1,0 +1,8 @@
+package dev.tssvett.schedule_bot.backend.exception;
+
+public class ScheduleBotException extends RuntimeException {
+
+    public ScheduleBotException(String message) {
+        super(message);
+    }
+}
