--- src/main/java/dev/tssvett/schedule_bot/backend/exception/StudentNotExistsException.java
@@ -1,8 +1,0 @@
-package dev.tssvett.schedule_bot.backend.exception;
-
-public class StudentNotExistsException extends RuntimeException {
-
-    public StudentNotExistsException(String message) {
-        super(message);
-    }
-}
