+++ src/main/java/dev/tssvett/schedule_bot/backend/exception/annotation/PostBeanProcessorException.java
@@ -1,0 +1,10 @@
+package dev.tssvett.schedule_bot.backend.exception.annotation;
+
+import dev.tssvett.schedule_bot.backend.exception.ScheduleBotException;
+
+public class PostBeanProcessorException extends ScheduleBotException {
+
+    public PostBeanProcessorException(String message) {
+        super(message);
+    }
+}
