+++ src/main/java/dev/tssvett/schedule_bot/backend/exception/database/GroupNotExistException.java
@@ -1,0 +1,9 @@
+package dev.tssvett.schedule_bot.backend.exception.database;
+
+import dev.tssvett.schedule_bot.backend.exception.ScheduleBotException;
+
+public class GroupNotExistException extends ScheduleBotException {
+    public GroupNotExistException(String message) {
+        super(message);
+    }
+}
