+++ src/main/java/dev/tssvett/schedule_bot/backend/exception/parse/ParseElementException.java
@@ -1,0 +1,10 @@
+package dev.tssvett.schedule_bot.backend.exception.parse;
+
+import dev.tssvett.schedule_bot.backend.exception.ScheduleBotException;
+
+public class ParseElementException extends ScheduleBotException {
+
+    public ParseElementException(String message) {
+        super(message);
+    }
+}
