+++ src/main/java/dev/tssvett/schedule_bot/backend/service/GroupService.java
@@ -1,7 +1,7 @@
 package dev.tssvett.schedule_bot.backend.service;
 
 import dev.tssvett.schedule_bot.backend.entity.Group;
-import dev.tssvett.schedule_bot.backend.exception.GroupNotExistException;
+import dev.tssvett.schedule_bot.backend.exception.database.GroupNotExistException;
 import dev.tssvett.schedule_bot.backend.repository.GroupRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
