+++ src/main/java/dev/tssvett/schedule_bot/backend/service/StudentService.java
@@ -4,8 +4,8 @@ import dev.tssvett.schedule_bot.backend.entity.Student;
 import dev.tssvett.schedule_bot.backend.entity.Faculty;
 import dev.tssvett.schedule_bot.backend.entity.Group;
 import dev.tssvett.schedule_bot.backend.entity.Notification;
-import dev.tssvett.schedule_bot.backend.exception.NotValidRegistrationStateException;
-import dev.tssvett.schedule_bot.backend.exception.StudentNotExistsException;
+import dev.tssvett.schedule_bot.backend.exception.registration.NotValidRegistrationStateException;
+import dev.tssvett.schedule_bot.backend.exception.database.StudentNotExistsException;
 import dev.tssvett.schedule_bot.backend.repository.NotificationRepository;
 import dev.tssvett.schedule_bot.backend.repository.StudentRepository;
 import dev.tssvett.schedule_bot.bot.enums.RegistrationState;
