+++ src/main/java/dev/tssvett/schedule_bot/bot/annotation/postprocessor/RegistrationRequiredPostProcessor.java
@@ -1,6 +1,6 @@
 package dev.tssvett.schedule_bot.bot.annotation.postprocessor;
 
-import dev.tssvett.schedule_bot.backend.exception.PostBeanProcessorException;
+import dev.tssvett.schedule_bot.backend.exception.annotation.PostBeanProcessorException;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.annotation.RegistrationRequired;
 import lombok.extern.slf4j.Slf4j;
