+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/faculty/FacultyKeyboardButton.java
@@ -2,7 +2,7 @@ package dev.tssvett.schedule_bot.bot.keyboard.impl.faculty;
 
 import dev.tssvett.schedule_bot.backend.entity.Student;
 import dev.tssvett.schedule_bot.backend.entity.Faculty;
-import dev.tssvett.schedule_bot.backend.exception.NotValidRegistrationStateException;
+import dev.tssvett.schedule_bot.backend.exception.registration.NotValidRegistrationStateException;
 import dev.tssvett.schedule_bot.backend.service.FacultyService;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.actions.keyboard.impl.details.CallbackDetails;
