+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/group/GroupKeyboardButton.java
@@ -1,7 +1,7 @@
 package dev.tssvett.schedule_bot.bot.keyboard.impl.group;
 
 import dev.tssvett.schedule_bot.backend.entity.Group;
-import dev.tssvett.schedule_bot.backend.exception.NotValidRegistrationStateException;
+import dev.tssvett.schedule_bot.backend.exception.registration.NotValidRegistrationStateException;
 import dev.tssvett.schedule_bot.backend.service.GroupService;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.actions.keyboard.impl.details.CallbackDetails;
