+++ src/main/java/dev/tssvett/schedule_bot/bot/keyboard/impl/reregister/ReRegistrateButton.java
@@ -1,6 +1,6 @@
 package dev.tssvett.schedule_bot.bot.keyboard.impl.reregister;
 
-import dev.tssvett.schedule_bot.backend.exception.NotValidRegistrationStateException;
+import dev.tssvett.schedule_bot.backend.exception.registration.NotValidRegistrationStateException;
 import dev.tssvett.schedule_bot.backend.service.StudentService;
 import dev.tssvett.schedule_bot.bot.actions.keyboard.impl.details.CallbackDetails;
 import dev.tssvett.schedule_bot.bot.formatter.message.MessageConstants;
